fix(export): require a true single parenthetical, and keep balanced parens in a url
Both review findings reproduced first, and both were real.

Data loss: the placeholder-section test checked only the first and last
character, so "(B.Sc.) Computer Science (2020)" satisfied it and the whole
EDUCATION section was silently deleted - measured, the section vanished from
the model. It now tracks depth and requires the opening paren to stay unclosed
until the final character, so a paragraph that merely starts and ends with
parens keeps its section.

Truncated links: the url regex stopped at every ")", so
"example.com/Function_(mathematics)" lost its closing paren and became a dead
link. The regex now keeps ")" and a balance check decides afterwards - a
closing paren survives only when an unclosed "(" inside the token is waiting
for it, so "(see example.com/a)." still trims correctly either way round.

The test module moved to model/adapter/tests.rs, matching model_docx.rs: the
new cases pushed adapter.rs past the R8 module cap.

Both guards mutation-checked - reverting either predicate fails its test.
